many changes for ver 0.7

+++ astwro/utils/CycleFile.py
@@ -5,12 +5,6 @@ class CycleFile(object):
     """
     Creates series of files with names containing counter, with symlink to newest one
     """
-    counter = 0
-    file = None
-    filename_formatter = '{}{:03d}{}'
-    symlink_formatter = '{}{}{}'
-    file_mode = 'w'
-    open_kwargs = None
 
     def __init__(self, path, basename, extension='', create_symlinks=True, symlink_suffix='_last', auto_close=True):
         """
@@ -28,11 +22,24 @@ class CycleFile(object):
         self.create_symlinks = create_symlinks
         self.symlink_suffix = symlink_suffix
         self.auto_close = auto_close
+        self.counter = 0
+        self.file = None
+        self.filename_formatter = '{}{:03d}{}'
+        self.symlink_formatter = '{}{}{}'
+        self.file_mode = 'w'
+        self.open_kwargs = None
+        self.filepath = None
         pass
+
+    def __enter__(self):
+        return self
+    def __exit__(self, type, value, traceback):
+        self.close()
+        return True
 
     def __del__(self):
         if self.auto_close and self.file is not None:
-            self.file.close()
+            self.close()
 
     def next_file(self, counter=None):
         if counter is None:
@@ -41,16 +48,27 @@ class CycleFile(object):
         if self.auto_close and self.file is not None:
             self.file.close()
             self.file = None
+        self._link()
         filename = self.filename_formatter.format(self.basename, counter, self.extension)
         filepath = os.path.join(self.path, filename)
         if self.open_kwargs is not None:
             self.file = open(filepath, self.file_mode, **self.open_kwargs)
         else:
             self.file = open(filepath, self.file_mode)
-        if self.create_symlinks:
+        self.filepath = filepath
+        return self.file
+
+
+    def close(self):
+        if self.file is not None:
+            self.file.close()
+        self._link()
+
+
+    def _link(self):
+        if self.create_symlinks and self.filepath:
             symlname = self.symlink_formatter.format(self.basename, self.symlink_suffix, self.extension)
             symlpath = os.path.join(self.path, symlname)
             if os.path.exists(symlpath):
                 os.remove(symlpath)
-            os.symlink(filename, symlpath)
-        return self.file
+            os.symlink(self.filepath, symlpath)
